Memperbaiki Controller Dengan Tabel KHS
+ Memperbaiki Eloquent Tabel KHS untuk Controler NilaiMahasiswa, KHS
+ Memperbaiki Controller Riwayat Absen
+ Memperbaiki Controller UbahAbsen
+ Menambahkan package laravel-qr-code
+ Menambahkan Qr-code untuk Absen
+ Mengubah Model DetailKrs, Khs, Krs, Nilai
+ Mengubah Migration Khs dan nilai
+ Menambahkan download krs
+ Menambahkan package dompdf

+++ database/migrations/2024_07_27_113902_create_khs_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
                 table:'krs',
                 indexName: 'nilai_id_krs'
             );
-            $table->integer('ips')->nullable();
+            $table->float('ips')->nullable();
             $table->timestamps();
         });
     }
